Update on "test Int8DynActInt4WeightLinear"
ghstack-comment-id: 2430970002

[ghstack-poisoned]
